% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substituting_formula.R
\name{formula.substituting_formula}
\alias{formula.substituting_formula}
\title{Convert a substituting_formula object into a regular formula.}
\usage{
\method{formula}{substituting_formula}(x, ...)
}
\arguments{
\item{x, object}{the substituting_formula object}

\item{...}{Ignored}
}
\value{
A formula with values substituted.
}
\description{
Convert a substituting_formula object into a regular formula.
}
