% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_side.R
\name{formula_side}
\alias{formula_side}
\alias{get_lhs}
\alias{get_rhs}
\title{Extract formula parts}
\usage{
get_lhs(x)

get_rhs(x)
}
\arguments{
\item{x}{A formula (or something that can be coerced to a formula) to extract
a part from}
}
\value{
The requested part of the formula as a name or call or \code{NULL} if
  it does not exist.
}
\description{
Extract formula parts
}
\section{Functions}{
\itemize{
\item \code{get_lhs}: Extract the left hand side (NULL for one-sided
formula).

\item \code{get_rhs}: Extract the right hand side.
}}

