% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substituting_formula.R
\name{formula_substituting_formula}
\alias{formula_substituting_formula}
\alias{formula.substituting_formula}
\alias{as.formula.substituting_formula}
\title{Convert a substituting_formula object into a regular formula.}
\usage{
\method{formula}{substituting_formula}(x, ...)

as.formula.substituting_formula(object, env = environment(object$base))
}
\arguments{
\item{x, object}{the substituting_formula object}

\item{...}{Ignored}

\item{env}{The environment for the resulting formula}
}
\value{
A formula with values substituted.
}
\description{
Convert a substituting_formula object into a regular formula.
}
