% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_parens.R
\name{simplify_parens}
\alias{simplify_parens}
\title{Remove extraneous parentheses from a formula.}
\usage{
simplify_parens(x)
}
\arguments{
\item{x}{The formula (or call) to simplify}
}
\value{
The simplified formula
}
\description{
Remove extraneous parentheses from a formula.
}
\examples{
simplify_parens(((a))~((b+c)))
}
